Implement topological sort
Implement the graph as a simple hashmap from vertices to adjacency
lists. Going to need a wrapper around it though, because right now it
can be in an illegal state, i.e. a vertex can be in an adjacency list
but not exist as a vertex with its own adjacency list. For consistency,
a vertex that exists but has 0 outdegree should be represented as an
entry in the hashmap with an empty adjacency list, rather than by simply
being absent.